PRODUCT VARIATION 5. DISPLAYING VARIATIONS INSIDE A CART
Adding a for loop inside the 2nd. try except block of add_cart function in carts.views, looping over the variations and adding them to the cart. if the cart item doesn't exist a new cart item with a quantity of 1 is created and linked to the current product and cart;removal of added to cart and view cart buttons from product_detail.html as they were interfering with adding multiple product varitions to the cart;
adding an if statement and for loop to cart.html: if there are any variations connected to the cart item for loop will run, looping over all the variations of the product rendering the category and value of each variation